==> 11_ExplorationsDataCoverage.R <==
##%######################################################%##
#                                                          #
####            Explorations: Data coverage             ####
#                                                          #
##%######################################################%##


# In this script, we explore the coverage of the data included within 
# various subsets of the data. 


# Charlie to complete.
